+++ changelog.md
@@ -1,3 +1,15 @@
+v0.3.8
+------
+
+- Renamed "MainWindow" to "NinjaCalc", closes #57.
+- Improved the image for Track Current (IPC-2152) calculator, closes #48.
+- Fixed bug where you couldn't see the app icon on Windows 10 task bar, closes #36.
+- Signed the application so warnings don't occur on install, closes #47.
+- Added "oz" as a unit for track thickness in the IPC-2221A calculator, closes #60.
+- Created a static class to hold unit conversion constants which are shared among calculators, closes #61.
+- Added info about the Visio calculator diagrams to the README, closes #33.
+- Added low pass RC filter calculator, closes #50.
+
 v0.3.7
 ------